+++ tests/bootstrap.php
@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 // report all errors
 error_reporting(-1);