+++ docs/CHANGELOG.md
@@ -1,8 +1,9 @@
-# Version 1.2.2 2018-10-18
+# Version 1.3.0 2018-10-18
 - Improve `XmlResourceRetriever\AbstractBaseRetriever::urlParts`
     - Remove deprecated useless constants `FILTER_FLAG_SCHEME_REQUIRED` or `FILTER_FLAG_HOST_REQUIRED`
     - Add docblocks
-- Add context features to `XmlResourceRetriever\Downloader\PhpDownloader` and tests
+- Add stream context features to `XmlResourceRetriever\Downloader\PhpDownloader` and tests
+- Add strict_types to all project
 
 # Version 1.2.1 2018-09-04
 - An XML resource that is an empty file is declared as not valid